requirements.txt
streamlit
python-dotenv
pandas
openpyxl
python-docx
pdfplumber
google-generativeai
